Publish images for each open pull request
CircleCI was previously only configured to publish images for pushes
to branches or tags.  While CI runs against open PRs, the publish jobs
that published images was excluded.

This PR updates the deploy jobs to detect when they are running
against a pull request.  When run against a PR, the images are pushed
to the `metallbdev` org instead of `metallb` and the image tag is
"pr-<ID>".

The dev docs on the web site are also updated to note how to pull an
image built for a PR.

Fixes #623